поместил класс OrderPanel в отдельный класс
